Moved and renamed the view pager indicator
