+++ WooCommerce/src/main/res/layout-sw600dp/fragment_login_prologue.xml
@@ -43,7 +43,7 @@
         app:layout_constraintTop_toBottomOf="@+id/image_prologue"
         tools:background="@color/blue_dark" />
 
-    <com.woocommerce.android.ui.login.LoginPrologueViewPagerIndicator
+    <com.woocommerce.android.widgets.WCViewPagerIndicator
         android:id="@+id/viewPagerIndicator"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
